Revise CacheControl to be able to bundle and forward to multiple caches
CacheIoTiming was moved out of the CacheControl interface. I/O statistics
and time budget is now global per ThreadGroup and can be obtained through
static methods of CacheIoTiming.

The number of priority levels in the IoTimeBudget is not fixed anymore. It
adapts when a new budget is set. Querying and using budget clamp specified
level to the current number of priority levels.